requirements.txt
django
djangorestframework
Celery
aiogram
dotenv
drf-yasg
django-cors-headers
djangorestframework-simplejwt